Implement compiler arguments
Added support for compiler arguments to g++ within the cell magic declaration. Any arguments after "--" will be passed to as arguments to g++, and anything before will be passed as arguments to the cpp magic. For example, you can use "%%cpp -t -- -O3 -lpthread" at the beginning of your cell. This will run "timeit" on the cpp magic, while passing the "-O3" and "-lpthread" arguments to g++.